-f on press back in result go to home -f remove unused files + add Home Screnn -f fix on finish flag game -f extract composable to files -f add default values
