+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/MainActivity.kt
@@ -10,11 +10,10 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import ar.edu.unlam.mobile.scaffolding.ui.screens.AdvancedGameScreen
+import ar.edu.unlam.mobile.scaffolding.ui.screens.GameAdvancedScreen
 import ar.edu.unlam.mobile.scaffolding.ui.screens.GameClassicResultScreen
 import ar.edu.unlam.mobile.scaffolding.ui.screens.GameClassicScreen
 import ar.edu.unlam.mobile.scaffolding.ui.screens.HomeScreen
@@ -48,24 +47,24 @@ fun MainScreen() {
     Scaffold { paddingValue ->
         // NavHost es el componente que funciona como contenedor de los otros componentes que
         // podrán ser destinos de navegación.
-        NavHost(navController = controller, startDestination = NavHostRouterPaths.GAME_CLASSIC.route) {
+        NavHost(navController = controller, startDestination = NavHostRouterPaths.HOME.route) {
             // composable es el componente que se usa para definir un destino de navegación.
             // Por parámetro recibe la ruta que se utilizará para navegar a dicho destino.
             composable(NavHostRouterPaths.HOME.route) {
                 // Home es el componente en sí que es el destino de navegación.
-                HomeScreen(modifier = Modifier.padding(paddingValue))
+                HomeScreen(modifier = Modifier.padding(paddingValue), controller)
             }
             composable(NavHostRouterPaths.GAME_CLASSIC.route) {
                 // Home es el componente en sí que es el destino de navegación.
-                GameClassicScreen(controller, viewModel = hiltViewModel())
+                GameClassicScreen(controller)
             }
             composable(NavHostRouterPaths.GAME_RESULT.route) {
                 // Home es el componente en sí que es el destino de navegación.
                 GameClassicResultScreen(controller)
             }
-            composable(NavHostRouterPaths.ADVANCED_GAME.route) {
+            composable(NavHostRouterPaths.GAME_ADVANCED.route) {
                 // Home es el componente en sí que es el destino de navegación.
-                AdvancedGameScreen(controller, viewModel)
+                GameAdvancedScreen(controller)
             }
         }
     }
