+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/NavHostRouterPaths.kt
@@ -4,5 +4,5 @@ enum class NavHostRouterPaths(val route: String) {
     HOME("home"),
     GAME_CLASSIC("GameClassic"),
     GAME_RESULT("GameResult"),
-    ADVANCED_GAME("AdvancedGame"),
+    GAME_ADVANCED("GameAdvanced"),
 }
