--- app/src/main/java/ar/edu/unlam/mobile/scaffolding/ui/components/Greeting.kt
@@ -1,16 +1,0 @@
-package ar.edu.unlam.mobile.scaffolding.ui.components
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-fun Greeting(
-    name: String,
-    modifier: Modifier = Modifier,
-) {
-    Text(
-        text = "Hola $name!",
-        modifier = modifier,
-    )
-}
